added intial tests
