+++ isis/tests/data/mgsImages/Testing/Temporary/CTestCostData.txt
@@ -1,0 +1,1 @@
+---
