+++ isis/tests/data/mgsImages/Testing/Temporary/LastTest.log
@@ -1,0 +1,3 @@
+Start testing: Feb 26 11:54 MST
+----------------------------------------------------------
+End testing: Feb 26 11:54 MST
